Implement Node-API for Hermes (#61)
This is an implementation of Node-API for Hermes.

### Why do we need this change?

Currently Hermes engine does not offer an ABI safe API that can be used on DLL boundary.
It is not a problem if Hermes code is embedded into an application which uses compatible build system. Though, it does not work well in scenarios when a third-party app does not use the same build system or the same programming language such as C#. The issue can be addressed by distributing Hermes as a DLL with an ABI-safe API.

While it is possible to implement Hermes-specific ABI-safe API, this PR follows a different approach: we adopt the Node-API from Node.JS project that could be considered as an industry standard in that area. It should allow writing native code that works with any JS engine that implement the Node-API. Imagine that developers could create a native module in the language of their choice that works in Node.js, Electron, and ReactNative applications. Since Node-API is already implemented for V8 JS engine, the Node-API for Hermes implementation makes us one step closer to such goal.

We already have the JSI implemented on top of the Node-API in ReactNative for Windows project and any existing ReactNative JSI-based code could continue to work without any changes. The Node-API is C-based API and ideally must never be used directly from C++ code that may throw exceptions. C++ developers should either use JSI or the Node-API C++ wrappers from Node.JS project. In future we must be able to add a .Net wrapper for Node-API to enable use of C# with the JS engines.

### Implementation details

The core of the implementation is in the `hermes_napi.cpp` file. It is done in the single file to match the JSI implementation in `hermes.cpp`.

These files are copied from Node.js sources code and represent the core Node-API:
- `js_native_api.h`
- `js_native_api_types.h`

The `js_native_ext_api.h` is copied from the `v8-jsi` project as the extensions to Node-API required to host JS engine and to implement JSI.

The implementation has required changes to Hermes code:
- Added new `FinalizableNativeConstructor` class inherited from `NativeConstructor` to allow defining classes in native code.
- `UTF8.h` and `UTF8.cpp` - added functions to simplify conversion from UTF16 to UTF8 where a target is an arbitrary buffer instead of `std::string`.
- `JSArrayBuffer.h` and `JSArrayBuffer.cpp` - added support for external buffers.
- `StorageProvider.cpp` - fix use of a wrong variable to suppress the unused variable message.
- Fixed TypedArray property implementation for detached buffers.

### Unit tests

The NAPI unit tests were adopted from the [microsoft/v8-jsi](https://github.com/microsoft/v8-jsi) project.
They have some modifications due to some Hermes vs V8 incompatibilities:
- Hermes does not support BigInt
- Hermes does not support sparce Arrays
- Different error messages

Hermes does not support the full ES6 specification yet, and the unit tests add the Babel transforms with the presets targeting Hermes engine for ReactNative.