Remove juno folder prior to build during CI to silence CG alerts. (#163)
There's an unsupported rust tool called juno under the folder unsupported, which checks in a cargo.lock file that is causing component governance alerts. The juno folder is not touched during build whatsoever, so to silence component governance, we remove that folder before executing build. If the juno tool does become supported at some point, we can revisit this.